% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdplm.R
\name{sdplm}
\alias{sdplm}
\title{Sparse Dirichlet process linear regression}
\usage{
sdplm(X, y, intercept = TRUE, tau.scale.prior = 1, sb.truncation = 25,
  nsave = 1000, thin = 5, burnin = 5000,
  .rng.seed = as.integer(Sys.time()))
}
\arguments{
\item{X}{Design matrix of predictors}

\item{y}{Vector of outcomes}

\item{intercept}{Whether or not to include a model intercept. This parameter is
intended to provide a similar interface to
\code{glmnet::\link[glmnet]{glmnet}},
but is not currently in used. My eventual intent is to allow
users to pass in a separate matrix of predictors, \code{Z},
with associated coefficients given uninformative priors rather than
the sparse Dirichlet process formulation}

\item{tau.scale.prior}{Hyperparameter for the global scale of the Dirichlet process atoms,
\eqn{\tau_0}. Default is 1}

\item{sb.truncation}{Truncation level for the stick breaking process/number of atoms
kept in memory. Default is 25}

\item{nsave}{MCMC parameter: number of posterior samples to return. Default is 1000}

\item{thin}{MCMC parameter: thinning factor so only 1 in \code{thin} posterior
samples are returned. Default is 5. The total number of MCMC iterations
is \code{burnin} + \code{nsave} * \code{thin}}

\item{burnin}{MCMC parameter: number of samples to discard from the start of the
chain}

\item{.rng.seed}{Random number generator seed for thread safety and consistency.
Default is set from the clock}
}
\value{
An S3 object of class \code{sdplm} containing posterior samples
of each parameter
}
\description{
Fits a linear model with a Horseshoe-centered Dirichlet
process prior on the regression coefficients using a blocked Gibbs
sampler. The prior is very flexible and results in approximately sparse
coefficient clustering from iteration to iteration. Each call to
\code{sdplm} results in one MCMC chain.
}
