% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{sparseDP}
\alias{sparseDP}
\alias{sparseDP-package}
\title{sparseDP: Sparse Dirichlet process regression}
\description{
A work in progress. Only Gaussian likelihood/linear regression implemented
at present (see \code{\link{sdplm}}). \code{sparseDP} is intended to
solve the homogeneity pursuit problem from a Bayesian regression perspective
by placing an approximately sparse Dirichlet process prior directly on
the regression coefficients.
}
